mdf type

+++ cell-core/src/main/java/com/lazycece/cell/core/infra/dal/po/CellRegistryPO.java
@@ -35,15 +35,15 @@ public class CellRegistryPO {
     /**
      * current value
      */
-    private Long value;
+    private Integer value;
     /**
      * min value
      */
-    private Long minValue;
+    private Integer minValue;
     /**
      * max value
      */
-    private Long maxValue;
+    private Integer maxValue;
     /**
      * the step, that the interval size of the value.
      */
@@ -73,27 +73,27 @@ public class CellRegistryPO {
         this.name = name;
     }
 
-    public Long getValue() {
+    public Integer getValue() {
         return value;
     }
 
-    public void setValue(Long value) {
+    public void setValue(Integer value) {
         this.value = value;
     }
 
-    public Long getMinValue() {
+    public Integer getMinValue() {
         return minValue;
     }
 
-    public void setMinValue(Long minValue) {
+    public void setMinValue(Integer minValue) {
         this.minValue = minValue;
     }
 
-    public Long getMaxValue() {
+    public Integer getMaxValue() {
         return maxValue;
     }
 
-    public void setMaxValue(Long maxValue) {
+    public void setMaxValue(Integer maxValue) {
         this.maxValue = maxValue;
     }
